+++ Code/AT_1Init.lua
@@ -3,7 +3,7 @@
 -- All rights reserved, duplication and modification prohibited.
 -- You may not copy it, package it, or claim it as your own.
 -- Created May 1st, 2019
--- Updated May 22nd, 2019
+-- Updated May 29th, 2019
 
 
 local lf_print = false -- Setup debug printing in local file
@@ -17,6 +17,7 @@ g_AT_Options = {
 	ATvoyageWaitTime    = 5,         -- Wait this amount of sols between voyages
 	ATrecallRadius      = true,      -- display recall radius on landed rocket
 	ATearlyDepartures   = true,      -- allow for earlier departures when voyages waiting
+	ATstripSpecialty    = true,      -- strip a tourists specialty upon arrival
 } -- g_AT_Options
 
 -- Save game fixup variables
@@ -275,6 +276,17 @@ function OnMsg.RocketLaunchFromEarth(rocket)
       local cargo = {}
 
       if #tourists > 0 then
+      	if g_AT_Options.ATstripSpecialty then
+      	  -- remove specializations from tourists
+      	  for i = 1, #tourists do
+      	  	if tourists[i].specialist ~= "none" then
+      	  		tourists[i].traits[tourists[i].specialist] = nil
+      	  		tourists[i].specialist = "none"
+      	  		tourists[i].traits.none = true
+      	  	end -- if tourists[i]
+      	  end -- for i
+      	end -- if g_AT_Options.ATstripSpecialty
+
         cargo[1] = {
           class = "Passengers",
           amount = count,